refactor: repoint imports from tools.processor to tools.hitl
Pause classes (ConfirmationPause / UserInputPause /
ExternalExecutionPause) are now exported from tools.hitl. Update
every internal import site:

- agent/agent.py: top-level import + two function-local imports
  for HITL handling and the user-input retry path.
- agent/pipeline/manager.py: TYPE_CHECKING import + the runtime
  import in the pause-handling branch.
- agent/pipeline/steps.py: pause import inside the model-request
  step.
- tools/user_input.py: function-local import of UserInputPause
  inside get_user_input — kept function-local to honour the
  hitl.py <-> user_input.py lazy-import contract.
- models/__init__.py: tool_manager.processor.extract_builtin_tools
  -> tool_manager.normalizer.extract_builtin_tools.

Also swap a handful of internal accessors:
- tool_manager.processor.registered_tools -> tool_manager.registry.registered_tools
- tool_manager.wrapped_tools -> tool_manager.registry.wrapped_tools
